requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
Flask==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
zope.interface==6.3
